# contracts/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/project-runs": {
      "get": {
        "operationId": "ProjectRunsController_listProjectRuns",
        "parameters": [
          {
            "name": "state",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "READY",
                "ACTIVE",
                "BLOCKED",
                "COMPLETED",
                "ARCHIVED"
              ],
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 50
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRunListResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}": {
      "get": {
        "operationId": "ProjectRunsController_getProjectRun",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRunProjectionDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}/tasks/{taskId}/start": {
      "post": {
        "operationId": "ProjectRunsController_start",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}/tasks/{taskId}/defer": {
      "post": {
        "operationId": "ProjectRunsController_defer",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}/tasks/{taskId}/block": {
      "post": {
        "operationId": "ProjectRunsController_block",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}/tasks/{taskId}/resume": {
      "post": {
        "operationId": "ProjectRunsController_resume",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}/tasks/{taskId}/verify": {
      "post": {
        "operationId": "ProjectRunsController_verify",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}/tasks/{taskId}/ai-help": {
      "post": {
        "operationId": "ProjectRunsController_aiHelp",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "taskId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRunAiHelpRequestDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRunAiHelpResponseDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}/archive": {
      "post": {
        "operationId": "ProjectRunsController_archive",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}/pull-request": {
      "post": {
        "operationId": "ProjectRunsController_bindPullRequest",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BindProjectRunPullRequestDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}/reverify": {
      "post": {
        "operationId": "ProjectRunsController_reverify",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "202": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}/publish": {
      "post": {
        "operationId": "ProjectRunsController_publish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
    "/api/project-runs/{id}/unpublish": {
      "post": {
        "operationId": "ProjectRunsController_unpublish",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "if-match",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "idempotency-key",
            "required": true,
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "project runs"
        ]
      }
    },
...
  },
  "info": {
    "title": "Jagalchi API",
    "description": "Jagalchi modular API contract",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ProjectRunListResponseDto": {
        "type": "object",
        "properties": {
          "items": {
            "maxItems": 50,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProjectRunProjectionDto"
            }
          },
          "nextCursor": {
            "type": "string",
            "nullable": true
          }
        },
        "required": [
          "items",
          "nextCursor"
        ],
        "additionalProperties": false
      },
      "ProjectRunAiHelpRequestDto": {
        "type": "object",
        "properties": {
          "question": {
            "type": "string",
            "maxLength": 2000
          }
        },
        "additionalProperties": false
      },
      "ProjectRunAiHelpProvenanceDto": {
        "type": "object",
        "properties": {
          "provider": {
            "type": "string",
            "maxLength": 80
          },
          "model": {
            "type": "string",
            "maxLength": 160
          },
          "promptVersion": {
            "type": "string",
            "maxLength": 80
          },
          "inputHash": {
            "type": "string",
            "pattern": "^[0-9a-f]{64}$"
          },
          "generatedAt": {
            "type": "string",
            "format": "date-time"
          }
        },
        "required": [
          "provider",
          "model",
          "promptVersion",
          "inputHash",
          "generatedAt"
        ],
        "additionalProperties": false
      },
      "ProjectRunAiHelpResponseDto": {
        "type": "object",
        "properties": {
          "guidance": {
            "type": "string",
            "minLength": 1,
            "maxLength": 4000
          },
          "provenance": {
            "$ref": "#/components/schemas/ProjectRunAiHelpProvenanceDto"
          }
        },
        "required": [
          "guidance",
          "provenance"
        ],
        "additionalProperties": false
      },
      "BindProjectRunPullRequestDto": {
        "type": "object",
        "properties": {
          "githubRepositoryId": {
            "type": "string",
            "pattern": "^[1-9]\\d{0,19}$"
          },
          "pullNumber": {
            "type": "integer",
            "minimum": 1
          }
        },
        "required": [
          "githubRepositoryId",
          "pullNumber"
        ],
        "additionalProperties": false
      },
...
    }
  }
}